194646: Don't reuse devices from a terminated session.
This is a targeted fix for
https://code.google.com/p/android/issues/detail?id=194646

The root cause is tracked by Bug 191552, a TODO that we didn't fix
before the release. There actually isn't much left to do to close the
original bug as well, but leaving it for a full review later.

This bug was fixed by a different CL for the 2.0 branch

Change-Id: If2c81a5eae89616711effd15d1c6f47d0851d1ed