Resource visibility clean-up
A small clean-up before we add proper handling of public and private
resources:
* Remove private R class package option - we will not be using it this
way
* Refactor ResourceAccessibility -> ResourceVisibility
* Refactor DEFAULT -> PRIVATE_XML_ONLY to better reflect what that
level of visibility represents.
* Add new res visibility: undefined - it represents a case where the
visibility was not defined at all; this should be only used in the case
when we are not generating the public and private R class.

Bug: 72735798
Test: existing tests
Change-Id: Ic2d538ab3a85cffad9aa5b3ebf9f981112481a61
(cherry picked from commit fe288490bc545d1e6ad145365e689510dd556835)