Allow Layout Editor to render after initial Blaze sync
The initial Blaze sync that happens when you re-open a project simply
loads the most recently cached model, even if the last sync from a
previous session failed and/or this model is out of date. For this
reason, and in order to remain consistent with the other sync status
restrictions on layout editor, ASWB will not let the layout editor
render until a successful sync has occurred during the current session.
This is really inconvenient for users who have a valid model and are
just switching between projects, as it forces them to sync every time
they open a project.

This CL is a temporary workaround that makes it so that ASWB considers
the initial Blaze sync to be sufficient for rendering the layout editor.
As this may lead to the user working with a stale model and not knowing
about it, future work should include either adding a warning banner
to tell the user they should sync to get an up-to-date model or
reintroducing the restriction but considering the startup sync to
be sufficient if the sync status was clean when the last session ended.

Bug: 78228400
Test: Ran BazelStartupSyncTest and NlEditorTest locally
      + tested manually

Change-Id: I224127d9bb509c0dea674e37642e2498718de27a
(cherry picked from commit caf5a0ea4d7e1bfec1d79282b14b7caf9c415ff3)