Not using layout editor configuration for nav
When a configuration is selected in the layout editor it's moved to the
front of the list in the static default ConfigurationManager, and thus
get picked up by the nav editor as well.
In this change we start using a separate ConfigurationManager for the
nav editor, so we keep using the default configuration priority order.
This means that if you create a landscape-specific navigation file it
will use a landscape configuration. TODO is to then actually draw the
destinations in landscape orientation.

Test: NavDesignSurfaceTest.testConfiguration()
Bug: 78589393
Change-Id: I57ebdc2e2ffe0dd435b526edacd2daa256b0b5ec
(cherry picked from commit d5a031417574a2fc935a668f763972357e535ca3)