Catch exceptions that are generated when Gradle build is cancelled
When Gradle build is cancelled an ExternalSystemException is generated, but
this type of Throwable does not contain a cause (it says why in the docs).
Insetad of using getCause, we need to check using getOriginalReason to see
if the task failed because it was cancelled.

Bug: 71492373
Test: Manualy confirmed that ExternalSystemException generated when a build
is canceled are processed correctly and that logs show the "Gradle execution
cancelled." message.

Change-Id: I52bf2c84191a76db6fb595b93beb691d6ce3a4bd
(cherry picked from commit 1e0a91549d065edb55a3a50dd6ba8c6e230519a5)