Replace all uses of SearchField with SearchTextField
SearchField was a UI customization of SearchTextField so that it would
look the same on GTK+ as on IntelliJ theme. However, this customization
crashes on the light theme on Windows 10. So we use the default
SearchTextField coming from JetBrains.

Bug: 110080380
Test: Updated UI tests
Change-Id: Ic0a04357b3312bba23c878a21fe7964d20d039f5
(cherry picked from commit 9adc5af5e16fa3cdf3ebde99d470d0e172316006)