Use annotations.zip from the platform with Android P DP5
As of Android P the platform ships with dynamically built
versions of annotations.zip instead of a prebuilt. We want
to use that instead of the extracted data bundled with
Studio.

Also customize import handler to rename @RecentlyNullable
to @Nullable, and ditto for @RecentlyNonNull.

Test: The separate annotations.zip file is tested by lint which
  loads it instead of the Studio one. Also verified manually.
Bug: 111667748
Change-Id: I6a3fb0ef9c36cf0a079580da47a512530b0c53af
(cherry picked from commit 0f76a058668320857a2b20f01d8435ded59772de)