Not rendering thumbnails 4x
Previously if we came in with multiple requests for the same thumbnail
at the same time we'd render all of them, which resulted in lots of
threads being created. Now we notice that this is happening and return
the same future for each.

Bug: 110231341
Test: ThumbnailManagerTest.testSimultaneousRequests
Change-Id: I7111c7aff5444ca40ada4ccb546c2e617e8f4852
(cherry picked from commit 29395c577ef3ec8f4d779c0cb97ba2565cd7afd3)