Don't check for library name mismatch in AarResourceRepositoryCache
The mismatch may come from the fact that the same AAR is used with or
without the model (e.g. if sync is triggered right after opening a
project) and there's no obvious way to recover the artifact address
(used when we have a model) from the Library name (used when we don't
have a model). This is fine, since for now we don't rely on exact
contents of the library name string, apart from checking if it's null.

Bug: 79942260
Test: manual, for 3.2 we just decided to remove the check, we'll work
  on keeping the same library name in 3.3.
Change-Id: I48044daec0e0a3c01602a51045d673d238164af4