Fix download link click logic.
On Windows the scrollbar width is included in the width of the
list of components in the palette. The logic for detecting a click
on the download image did not take this into account.
Fix the code such that this works on all platforms.

Bug: 111124139
Test: Manual for now, will attempt to add unit test later.
Change-Id: Ia8e0c47c560e53e9d16d3dadbf09cb076d310a1c
(cherry picked from commit 2006b19583f2b937ebed8f1ebfb5141107df28ea)