Check that artifacts can be resolved before adding them to result list
... when searching for libraries on the PSD.

When a support library is being looked for, we check that the same
version exists using a filter, however, if that results in an empty
list, the generated coordinate will now have a version, causing a null
artifact when calling Artifact.fromCoordinate. This will cause issues
when sorting and showing the list of available libraries.

With this change we first check that the coordinates resolve to a non
null artifact before adding it to the result.

Bug: 114329699
Test: Manually check that searching for 'support' does not cause an
exception.

Change-Id: I66073f58c1dce0c2e3403d8c9c3479ed83e66551
(cherry picked from commit 8f27dd72e6c63f5be08a1a7784b9b2a66d9d4e40)