Fix generated XML for TabLayout.
The XML generated for TabLayout accidentally included the
attributes: layout_width & layout_height twice.
This caused a TextView component with the invalid XML to be
created in the designer instead of a TabLayout component when it
was dragged from the palette.
Also: wrong item tag used for androidx.

Change-Id: Ia8c269eff8b388d6c7cec23a7c634763ebbfd6b6
Fixes: 118902963
Test: Added unit test.
(cherry picked from commit 57688bc6ca9aa5442ca1bda072848d19df1c9161)