Set the row height of property tables up front.
Before this change the row height was calculated based on the max
height of the first 5 rows in the table.
It turns out that on mac a CheckBox is rendered taller than
required, which made all the rows taller than necessary.

After this change CheckBox controls in tables will be slightly
squeezed, but is is not vivible with the human eye.

Bug: 122196629
Test: NA
Change-Id: I75da1820f476df2b9579ec5c133891e4b2a16255