Fix satisfaction dialog (currently not showing at all)
The dialog was not showing as an exception was thrown
due to the dialog being created off the UI thread.

This fix uses IdeEventQueue.addIdleListener
to ensure that the dialog is launched from the UI thread,
but at a time that the user is not actively busy coding.

Bug: none
Test: manual
Change-Id: Ic1a6b9af1a1f390981abf96f0b270993ba0a6297
(cherry picked from commit 0a96d0de9ce4753ad55a4758c5fd8597971196ad)