Do not request focus 2 different places.
A PropertyTextFieldWithLeftButton was requesting focus on the
TextField and the left button when focus was requested on the
model. Only the TextField should get it.

For the resource picker the focus used to go back to the table.
Instead force the focus back on the JComponent that caused the
event i.e. the icon. This works because the resource picker is
a modal dialog.

Change-Id: Ifa3ea5143b4bad0b020e4b768e43dcfa434bb6aa
Fixes: 122532254
Test: Manual tests added in a later CL.