Refactor color action to take advantage of data context.
Determine the active property from the data context for the
action button. Take advantage of this by turning
OpenResourceManagerAction and ColorSelectionAction into
singletons.

Use the AnAction registration for keys that activate the color
button actions for the flag editor and left button editor.

Minor: display the popup for color and flags arrow in the middle of
the left button icon.

Bug: 121286697
Test: NA
Change-Id: I18d118fe888452e2dee47a34f8b03a9d98df7314