Fix for bundle signing dialog export private key.
Only check to see if a destination is set if export private key is set.
Otherwise, the dialog can be left empty. In fact, the field is hidden if
the export private key is unchecked, so user cannot fill it in.

bug:122796456
Test: added unit test for dialog for this issue
Change-Id: I541e0aae1fd1a77c9a0faf65ed88d4f1fbf4310c
(cherry picked from commit 0144e5b07348e87f27daac342289bcc06c13f423)