Fix can't zoom-in/out in Layout Editor when scale is closed
When using Mac with Retina display, it is not able to zoom-in
Layout Editor if the current zoom-in value is 1% differences
to next built-in zooming point. For example, it is not able
to zoom-in from 24% to 25%, from 66% to 67%, and vice versa.

The reason is there has a minimal gap when setting scale. If
the differences between original scale and new scale is less
than the gap 0.5%, we don't change the scaling. This gap
should be divide by scaling factor since the extact scale
value is divide by scaling factor as well.

Fixes: 122881600
Test: Updated DesignSurfaceTest
Change-Id: I0d6286cb14130288056a864ac11eb187be0022fc