Cancel the popup when a default margin is set.
The margin popup used to hide the window of the popup. This was
causing problems for other popups since the code that is supposed
to automatically close the later popup determines the popup should
not be closed before the first popup is closed.

Change-Id: I0009e98ccf9d98a38fa2ec2ade739977bd1ee746
Fixes: 123071296
Test: N/A
(cherry picked from commit 223fdc2ca0f8421d867640b826342d6d37ff60cc)