Track data binding enabled metric
This is the first of many CLs to track data binding usage metrics. This logs whether data binding is enabled in build.gradle file and defines a service interface that is implented with a dummy implementatio in core, and overriden with an actual implementation in data binding.

TODO(b/123721754): data binding is enabled on a per module basis, and so should be tracked that way. Currently, we consider a project to be enabled if any of its module is enabled.

Bug: 123029035
Test: DataBindingDummyTrackerTest and DataBindingTrackerTest.
Change-Id: I3378c24bc32dd5f731fd72e193f13eb5f806f4b9