Setup project in background after changing variant
When a user changes the selected variant to another one that is already
sync we do not do a sync, instead, we used the previous result from
cache and from it we setup the project. We are currently doing all
project setup in a single write action, this causes the UI to be
unresponsive during project setup, this can be multiple seconds for
large project.

This changes moves project setup to a background task and inside of it
it does multiple write actions, instead of a single one. This improves
UI responsiveness.

Bug: 123936667
Test: Manualy verified that UI is responsive with a project with 130
modules after this change. Current tests on BuildVariantUpdaterTest
check cached sync and created a couple of tests for BathUpdatesUtil

Change-Id: Iaf6b2ca8ff713582e8e05193b4c3ccf845c26633
(cherry picked from commit 98f8487c31c0d98cc0aa143a7f2bd7df13f81867)