Fix auto variant selector to properly handle ABIs.
This CL does not change anything related to the UI layout. It just
fixes the bugs in the "Build Variants" window that were introduced
with the 3.3 release.

When the user changes the build variant of a non-native module:

Before this CL: Dependent native variants were not properly selected
if the newly selected variant has already been sync'ed before.

After this CL: The same variant is selected. The ABI is preserved.

Bug: 122875116
Test: Manually tested.
Change-Id: Icdd6b2498cff5caf290db6a6edb6a13c8bb727af
Signed-off-by: Emre Kultursay <emrekultursay@google.com>
(cherry picked from commit acf127c18955738ed15d8e2a8522c44a0296701d)