Do not update property values when selection is changed.
The properties panel should not update the values when we know
there is a pending selection change.
The selection change is handled on a background thread and usually
takes longer than a value change. This could cause the property
panel to handle a value change on a deleted component.

Fixes: 124116596
Test: Added unit test.
Change-Id: I9c97857751de79ab4c9672056282bc89d72bd3fb
(cherry picked from commit 12084c379a7cd09b7e89b903a47428e12b915076)