Assume agent unattachable after certain retry threshold
This change adds a max duration (10-sec) to how long the profilers
wait for an agent to attached. If the agent status remains unspecified
after such duration, the profilers assumes that that agent is unattachable
so that the UI does not get stuck in the loading screen.

Cherry picked from a66f3511a0e2201b845d8efa5e4eac2cc9712895

Bug: 124254814
Test: StudioProfilersTest#testAgentUnattachableAfterMaxRetries

Change-Id: I4c063d4f8a57ff74208fcd7fffc1f2d1dde452ee