[android tests] don't register Android SDK twice in AndroidManifestDomTest
Android SDK is registered in Sdks.createLatestAndroidSdk, and if we register it in the test we'll get two identical instances in ProjectJdkTable. And during tearDown when removeJdk was called for a first time, it disposed ProjectJdkImpl instance and fired jdkRemoved event, and RootIndex tries to rebuild itself and accessed content of disposed ProjectJdkImpl instance (because the second instance is still present in ProjectJdkTable).

(cherry picked from commit 6d95064725d7abae9f4bee7e980958d22dc84648)

Change-Id: I6d95064725d7abae9f4bee7e980958d22dc84648