[Gradle Dsl] Start using InterpolatedText for String injections.
We now support setting properties to InterpolatedText values, and
getting model raw and resolvable values of INTERPOLATED type.

Also, start reflecting this new change on the PSD side: Interpolated
values aren't considered as String anymore, and when requesting
ResolvedPropertyModel.asString() for a interpolated property, we get the
resolved value of the property (i.e. with all the string injections
resolved).

Bug: 148283067
Test: updated.
Change-Id: Ib8442afa2cd7ff87b837dea943fe75ad5ee0c1fc