Add support for customizing class transforms
This CL enables RenderTask/LayoutlibSceneManager users to provide new
transforms that can be applied.

The main use case will be, for now, to apply the live literals
transform, only when enabled, and not all the time. The live literals
transform applies to every constant read so, for animations, there is a
cost that we can avoid if not needed.
On top of that, this infrastructure allows avoiding transformations that
can modify the user code behaviour when not needed.

Bug: N/A
Test: Existing tests pass. Will be tested with Live Literals.
Change-Id: Ia7fac1544921fb2bf43d7ce7ea550b211439f172