[Dsl] Support freeCompilerArgs for kotlinOptions
This is needed to Do The Right Thing when supporting kotlin compiler
options.

The (generated) interface, in KotlinCommonToolOptions.kt in the
kotlin-gradle-plugin-api source code, defines this as a var of a
List<String>.  We do not yet have a LIST type (only MUTABLE_LIST), so
for now our best representation is that this is an UNSPECIFIED_FOR_NOW
property, with VAR (writeable) semantics in both Groovy and
KotlinScript.

Bug: 159160295
Test: added; existing tests strengthened
Change-Id: Idb740bd568efa84e791f005fd3bef3183a084401