Refactor AppInspectionApiServices
This CL is preparation for future bundling changes.

There will be a new API called "GetLibraryVersions" and it is a process-level call (not pertaining to individual inspectors). This API will be used by the IDE module to determine which inspector tabs can be launched. The most natural place to put it is in AppInspectionTarget, however AppInspectionTarget will have to be made public. In order to continue hiding functions that shouldn't be exposed publicly, this change makes AppInspectionTarget an abstract class so existing functions can be made internal.

In addition, AppInspectorLauncher was removed because in retrospect it added an extra unnecessary layer on top of the existing API. Now, its functionality is moved to AppInspectionApiServices.launchInspector.

Test: N/A - this is a refactoring CL.
Bug: 165319681
Change-Id: I2e9f92f2c4037cda7386bd073ac0f8bd4914cf78