Set the actual property via the color picker
When a new property is added in the declared properties section of
the layout editor, we used to set the value through the value
property of the NeleNewProperty instance.
The NeleNewProperty instance however is reclaimed for the next
new property to be added. This caused the color picker to only be
able to set the property once. The other changes would be ignored.
Fix this by going through the actual property instance.

Also the color picker would set an initial value of white if no
initial value was specified.
Fix this by passing a fake new color of white.

Fixes: 144379473
Test: Unit tests added.
Change-Id: Id89761a17a6f52aaf1ed0b531e38b4a33c56511a