Allow lower versions of Java for Gradle JDK
... than the version used for Android Studio.

When IdeJdks.getJdk is called we use ProjectJdkTable information
looking, for a compatible JDK in this order:

- JDK stored in Android SDK's additional data
- JDK's stored directly in the ProjectJdkTable
- Embedded JDK.

JDK compatibility is checked with IdeSdks#isJdkCompatible but this
method considers a JDK compatible only if the version is greater or
equal to the one running by Studio, causing JDK 8 to not be compatible
and thus not used, forcing the embedded  JDK to be used.

Also, when a JDK location is selected in the PSD we store it in the
ProjectJdkTable by looking all the Android Sdks and updating their
additional data. But this dialog at the same time saves the SDK path,
but in order to do so it removes all existing SDKs and creates new
ones without passing the selected JDK, causing the SDKS to use the
embedded JDK.

This CL changes IdeSdks#isJdkCompatible to return true for JDKs 8 or
above and also uses the selected JDK when recreating the Android Sdks.

Bug: 166476070
Test: Added some tests in IdeSdksAndroidTest to check that the embedded
Jdk 8 can be used and is compatible. More integration tests will be
added in b/167414916.

Change-Id: I95185982713e42423badf60503abb685af3a9e82