Propagate variant switching from app to feature modules
When switching the build variant of an Android app it is expected that
build variants of any feature modules are switched accordingly.
However, the default build variant propagation mechanism does not work
in this case since the main app does not depend on the feature modules.
This change introduces special support for app-to-feature-module
variant switching propagation.

Bug: 159377709
Test: BuildVariantsIntegrationTest
Change-Id: If5afff430582cd3fe01c6567f33f2be0c1e83b7a