Use layoutlib native for Studio unbundled
When running Studio from IntelliJ, the two layoutlib plugin system does
not work because the plugin classloading is done differently in that
case. So we need to only depend on the one layoutlib version we want to
use.
This removes the dependency on the standard version and keep the
dependency on the native layoutlib module.

Fixes: 169838154
Test: manual check
Change-Id: I5e02b12044dc76a8a17e50312332f0d6ce4ad7cc