Revert Change Ib78af2d6
Sorry, I didn't catch it in the review. However, after I tested it more thoroughly I found issue where the scrollbar is missing when there is no sufficient space. Also, each table column has a preferable width property and the new minColumnWidth is smaller than most of them. I think it would be nice to keep the min width small so that a user may shrink the column as they needed.

Change-Id: I9f2573ba6f5560f92d32da3aca3e0e70ae81f14f