Add output parser for instant execution report
When there are instant execution errors a report is generated and an url
to it is included in the output. The GradleIssueData in the reported
error does not include the whole text and in partiuclar does not have
the url to the report. Because of this we cannot add a link to it in the
same node as the error.

This change creates a new BuildIssueEvent when an Instant execution
report is detected, with a link to the generated report, so users can
see it more directly, without looking for it in the whole output.

Test: InstantExecutionReportParserTest.kt
Bug: 156552742
Change-Id: Ia65d92dac289395ab0992dbba244ba4e5ef85f18