Propagate variant switching in the breadth first manner
This change makes a difference in the case of variant selection
conflicts only. The expected result is undefined in this case and
with this change we will prefer variants selected by a shorter path.

This change is required to make the structure of the code more suitable
for parallelization with the new experimental Gradle API.

Bug: n/a
Test: n/a
Change-Id: I86e418b9d468be0124b47ebb239972fca646b53f