Allow live literals to cast to other types
When using live literals, the user, while typing might accidentally
change the type of the constant. This will cause the const value to
change type. For example, a color, usually a long 0xAAAAAAAA, when
removing enough digits, could become an int.
We need to make sure that we allow the new type to be casted up back to
long.

This code handles gracefully those cases allowing for lost precision in
the conversion.

If there is not valid conversion, we fall back to return the default
value.

Fixes: 171026149
Test: Adds tests for the valid conversions
Change-Id: Ibaad90e495159517b659c7b828fe9417a882dd08