[lint] fix nullable args to non-nullable function
The signature of GradleDetector.checkDslPropertyAssignment changed,
and then changed back (see the linked bug); in the rollback, this case
in LintIdeGradleVisitor was missed.  Fix the handling of the
qualifierExpression for top-level reference expressions with one
argument, and introduce additional defensiveness for the name
of the potential property (though I have never encountered a case
where that is null, the annotations imply it could be).

Bug: 170026127
Test: added
Change-Id: Ic82cd46e869d7230c67536a2e406535b7eb8f6b7
(cherry picked from commit 83837fa47f7d39015f0c3351b9599a63be4cd573)