Disable safe args before Kotlin 1.4.20 plugin release
It turns out that safe args IDE support for Kotlin relies on Kotlin
plugin APIs that are going to break once the Kotlin 1.4.20 plugin
is released in the near future, due to binary incompatibilities.

To get ahead of this, we're disabling the feature for now, and
we'll re-enable it once we fix the compatibility issue.

Bug: 173552268
Test: N/A
Change-Id: Ic169ce18b47e9c4114e2fc4a4254868fe86c252d