Create new pipeline package
... and group `legacy` and `transport` classes underneath it.
We are also planning a third `appinspection` subpackage, coming soon.

As part of this change, we also rename the `Default` prefixed
implementation classes to `Transport`, since we'll soon be adding
yet another implementation, and the naming will help set them apart.

No logic was modified in this commit! A few classes were renamed, and
the rest were moved.

Bug: 173745851
Test: N/A
Change-Id: I27d4abe352908f473ddf643051039a8089da9874