Use correct inspector jar
The resolved jar wasn't used to launch inspector. Instead app inspection
was using the jar bundled with Studio. And once the jar was removed when
bundling went live, this error started happening.

The issue wasn't noticeable in development because we use a dev version
of WM inspector that is located at a different path (and this path
exists).

Test: updated AppInspectionViewTest.
Fixes: 174852747
Change-Id: I8ac3e3379b44c44067e081f9ee3cf5a9854a068d