Optimize `MemoryClasifierView.findSmallestSuperSetNode`
The new `MemoryClassifieriew.findSmallestSuperSetNode` implementation
iteratively finds a smaller node that's still a superset of the target
instances. The search finishes when it looks at a node that either has
(1) immediate instances intersecting with the target set, or (2) 2+ children
indirectly intersecting with the target set.

The new implementation is more optimized by replacing all but the first
calls to `ClassifierSet.isSuperSetOf` with cheaper checks for overlapping
between sets.

Fixes: 173037220
Test: existing
Change-Id: I329bcca00ac9bbecc62c059d5c622896347ad799