Remove isImage call in ResourceNotificationManager
The isImage call might trigger a content read in the UI thread. Since
this runs for many events the benefit is probably not worth the cost of
accessing all the files.
We will just rely on the folder to notify image notifications since
everything in those folders should be images anyway.

Fixes: 173129276
Test: Existing tests pass
Change-Id: Ic89fdfcb56aa1a506902fd77a81da827823115e2