[New Module Wizard] Show main module for "Import..."
The UI to show the main imported module was
removed in Change-Id Iaccae58e80a21289374a86b2270f84d7e0204cc8
by mistake. This CL re-introduces that UI.
It also fixes the UI layout, to make it more similar
to other "New Module" Steps and makes sure the
main panel is re-validated when the UI changes.

Fixes: 174850419
Test: Visual inspection and SourceToGradleModuleStepTest
Change-Id: I9d17493df77234b989e46f5fe1f815b5423282fc