Improve the viewport position after zoominig in Design Tools
This CL changes some viewport calculation ways when zooming.

In Compose and Validaition Tool, do not always reset viewport position
to center when zooming by buttons or shortcuts. Now they use the
current viewport position to decide the result viewport (see below).

In Compose, Validation Tool, and Layout Editor, use the relative
position to decide the new viewport position after zooming. This
makes the new viewport position is more close to the zooming center.

This CL also fixes a issue in Layout Editor. When zoom level is high,
zooming twice may set the viewport to weird position. (A simple
reproducing way is zoom a layout to 150% and scroll to the end of
right-bottom side. Then pressing zoom button twice to zoom to 300%)

The SceneView size of Navigation Editor is same as viewport size, so
using old calcuatioon way is better. Thus this CL doesn't effect
Navigation Editor.

Fixes: 168560154
Test: manually(UI related)
Change-Id: I33c5bd5b6d182a23002ec2eb7d4cf2020c4705b7