Remove Gradle configurations producers from ignoredProducers list.
Prior enabling Gradle test runner as the default unit tests runner in
AS, we used to ignore all the Gradle runConfiguration producers.
, and the list is
saved in the project in runConfigurations.xml file.
If a user open such a project in the new AS that uses Gradle test
runner, the unit tests will be run through JUnit as AndroidJUnit is
disabled elready.
This change removes the Gradle producers from the list of
ignoredProducers, which would allow users run unit tests through Gradle
for existing projects.

Bug: 176167205
Test: verified manually
Change-Id: Ief381e231544479589ba3cae0949ac6154890f95