Use ActionManager to create ActionToolbar in AttachedToolWindow
With the merge of IJ 203.3645.34 a change was made to make all
actions in the AttachedToolWindow DumbAware. That change caused the
action buttons to be invisible in IJ with Studio plugin.

To support DumbAware actions we need to use ActionToolbarImpl which
we did not do before this change.

This change also allows us to use a DropDownAction as an additional
action inside a tool window header.

Also: remove warnings from the test.

Bug: 161154345
Test: Update AttachedToolWindowTest
Change-Id: Icbecf70709ab7c2ef6e153627be04da1a7ac50ce