Register @ChecksSdkApiLevel inspection and annotation quickfix support
This CL registers the new @ChecksSdkApiLevel detector as an inspection.
It also hooks up support for lint quickfixes which request for an
element to be annotated (which the new @ChecksSdkApiLevel check does).
A couple of older detectors were also switch over to the new mechanism,
so delete the custom handling of the quickfixes here.

Test: New and existing unit tests
Bug: None
Change-Id: I44db8c2c36ccecea660b64fad86d51d6a4e0b49e