Prevent changing table selection when click is outside of table rows
When the rows of a table don't fill the table's height,
it's possible for the user to click on the table but not
on any row.

This change prevents changing the tables selection when
the right click is not on a valid row.

Test: TableViewImplTest.kt
Bug: 176528789
Change-Id: I28b9b5ab9e68c22da148c04697cd805a2921ba9a
(cherry picked from commit 809a9e80b4394d00e7cb731eb3aad5163c9f4922)