Remove transport references from the resource package
This commit introduces a few new classes in the resource package that
represent data needed to fetch Android resource classes in Studio.
Before, this package was using the transport proto definitions directly,
but as we are migrating away from that framework, taking ownership of
our own copy of these structures helps decoupling.

As additional benefits, we can now add useful code docs to those
classes, as well as useful utility methods as well.

Bug: 170895806
Test: Covered by existing
Change-Id: Ie324edbba2d1507b08a63b81c5dab2cb60ef3ad9