[sync] Run file hash updates mostly on the background thread
The file hash updates scheduled by a startup activity to track whether a
gradle build file has been modified (and hence that the sync state is
out of date) was scheduled to run on the EDT.  Significant disk IO was
performed on background (dedicated) IO threads, but those threads come
from a limited pool and could be starved, for example if another project
is updating icons based on background investigation of directories
triggered by an Open... dialog.  At that point, the file hash thread
could make no progress, which would manifest as a freeze.

Instead, this change arranges to schedule the file hash computation to
run on a pooled thread, and explicitly request the read lock everywhere
it is needed.

Bug: 177796351
Test: existing unit tests, limited manual testing
Change-Id: Ibb4015e439ab245a9595942c089dc5d65487dc55