Refactor view inspector cache handling
This new version doesn't rely on generations but instead on the fact
that response / event ordering is guaranteed in the new app inspection
system.

We accomplish this by moving the view property cache inside the view
inspector client - this lets us update the state of the cache directly
as new events come in from the device.

In this initial CL, this may feels like a somewhat arbitrary
simplification, but its benefits will become more important in a followup
change that introduces Compose support (since the Compose API won't have
generations inside it).

Bug: 170895806
Test: Will be done later (b/177231212)
Change-Id: If509e9cc125505b1d516c8ebf4311422a2c3d09f