Polish the Select Multiple Devices dialog
Return the single default target from
SelectMultipleDevicesDialogTableModel::getTargets if the device is
running or if it has no snapshots. If a device is running it won't be
booted, so we don't need to display all the boot options for it. We're
only using the default target to keep track of the device key. If it has
no snapshots it will always be quick booted, so we don't need to return
ColdBootTarget.

Do not display the boot option in the dialog if the device is running
nor if it has no snapshots

Make Device::getText return String instead of Optional<String>

Bug: 120497509
Test: SelectMultipleDevicesDialogTableModelRowTest
Test: UpdaterTest
Test: Existing tests
Change-Id: I6e96501434d2d492bcdfe973dff620bf9d80bae8